requirements.txt
1.pandas==1.3.5
2.numpy==1.26.4
3.matplotlib==3.8.3
4.seaborn==0.13.2
5.scikit-learn==1.4.2
6.streamlit==1.34.0
